Moved OsmTileLoader to "tilesource" package
